created combined reducer
relates to #45
