+++ src/client/reducers/index.js
@@ -1,58 +1,9 @@
-/*
- * The reducer takes care of our data
- * Using actions, we can change our application state
- * To add a new action, add it to the switch statement in the homeReducer function
- *
- * Example:
- * case YOUR_ACTION_CONSTANT:
- *   return assign({}, state, {
- *       stateVariable: action.var
- *   });
- */
+import {combineReducers} from 'redux';
 
-import {
-    CHANGE_FORM,
-    SET_AUTH,
-    SENDING_REQUEST,
-    SET_ERROR_MESSAGE
-} from '../constants/AppConstants';
+import home from './home';
+import studentDashboard from './search_jobs_reducer.js';
 
-const assign = Object.assign;
-import beevrAPI from '../utils/beevrAPI.js';
-
-const initialState = {
-    formState: {
-        username: '',
-        password: ''
-    },
-    currentlySending: false,
-    loggedIn: beevrAPI.loggedIn(),
-    errorMessage: ''
-};
-
-// Takes care of changing the application state
-export default function homeReducer(state = initialState, action) {
-    switch (action.type) {
-    case CHANGE_FORM:
-        return assign({}, state, {
-            formState: action.newState
-        });
-        break;
-    case SET_AUTH:
-        return assign({}, state, {
-            loggedIn: action.newState
-        });
-        break;
-    case SENDING_REQUEST:
-        return assign({}, state, {
-            currentlySending: action.sending
-        });
-        break;
-    case SET_ERROR_MESSAGE:
-        return assign({}, state, {
-            errorMessage: action.message
-        });
-    default:
-        return state;
-    }
-}
+export default combineReducers({
+    home,
+    studentDashboard
+});
